requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
distlib==0.3.1
Django==3.2.3
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==19.3.0
importlib-metadata==4.0.1
importlib-resources==5.1.3
mysqlclient==2.0.3
pipenv==2020.11.15
psycopg2-binary==2.8.6
Pygments==2.5.2
PyMySQL==1.0.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
virtualenv==20.4.6
virtualenv-clone==0.5.4
zipp==3.4.1
drf_yasg
